+++ cpp/libs/opendnp3/src/opendnp3/outstation/EventBufferFacade.h
@@ -38,12 +38,12 @@ struct EventBufferFacade
 public:
 
 	EventBufferFacade(	   
-		openpal::LinkedListAdapter<SOERecord, uint16_t> sequenceOfEvents_,
-		openpal::StackAdapter<openpal::ListNode<SOERecord>*, uint16_t> selectedEvents_
+		openpal::LinkedListAdapter<SOERecord, uint32_t> sequenceOfEvents_,
+		openpal::StackAdapter<openpal::ListNode<SOERecord>*, uint32_t> selectedEvents_
 	);
 	
-	openpal::LinkedListAdapter<SOERecord, uint16_t> sequenceOfEvents;
-	openpal::StackAdapter<openpal::ListNode<SOERecord>*, uint16_t> selectedEvents;
+	openpal::LinkedListAdapter<SOERecord, uint32_t> sequenceOfEvents;
+	openpal::StackAdapter<openpal::ListNode<SOERecord>*, uint32_t> selectedEvents;
 };
